Move BagVerifierError out as a separate object
This means we'll be able to reference it elsewhere, for example in tests
for the standalone traits.
