+++ bag_verifier/src/main/scala/uk/ac/wellcome/platform/archive/bagverifier/verify/steps/VerifyChecksumAndSize.scala
@@ -2,20 +2,21 @@ package uk.ac.wellcome.platform.archive.bagverifier.verify.steps
 
 import uk.ac.wellcome.platform.archive.bagverifier.fixity.{FixityChecker, FixityListChecker, FixityListResult}
 import uk.ac.wellcome.platform.archive.bagverifier.fixity.bag.BagExpectedFixity
+import uk.ac.wellcome.platform.archive.bagverifier.models.BagVerifierError
 import uk.ac.wellcome.platform.archive.common.bagit.models.Bag
 import uk.ac.wellcome.platform.archive.common.storage.Resolvable
 import uk.ac.wellcome.storage.{ObjectLocation, ObjectLocationPrefix}
 
 import scala.util.{Failure, Success, Try}
 
-trait VerifyChecksumAndSize extends Step {
+trait VerifyChecksumAndSize {
   implicit val resolvable: Resolvable[ObjectLocation]
   implicit val fixityChecker: FixityChecker
 
   def verifyChecksumAndSize(
     root: ObjectLocationPrefix,
     bag: Bag
-  ): InternalResult[FixityListResult] = {
+  ): Either[BagVerifierError, FixityListResult] = {
     implicit val bagExpectedFixity: BagExpectedFixity =
       new BagExpectedFixity(root.asLocation())
